Cleaner fullrestart, and we now have a maximum limit on nextSimplify
Basically, we now abort both on full restart need and on simplify need,
PLUS we simplify always, every 60'000 restart AT MOST. Before there was
no cap, and the limit was determined in a geometrical fashion

+++ Solver/Gaussian.cpp
@@ -806,7 +806,7 @@ void Gaussian::disable_if_necessary()
             disabled = true;
 }
 
-llbool Gaussian::find_truths(vec<Lit>& learnt_clause, int& conflictC)
+llbool Gaussian::find_truths(vec<Lit>& learnt_clause, uint64_t& conflictC)
 {
     Clause* confl;
